20260719_1101_RES_Phase22-1_ThumbnailEngine task1

+++ src/Renderer_TrackDrum.h
@@ -26,7 +26,7 @@ public:
     const std::array<DrumSlot, 2>& GetDrumSlots() const { return m_drumSlots; }
     int GetCurrentDrumSlotIndex() const { return m_currentDrumSlotIndex; }
     float GetDrumRelativePosition() const { return m_drumRelativePosition; }
-    size_t GetAnimatingTargetIndex() const { return m_animatingTargetIndex; }
+    int GetAnimatingTargetIndex() const { return m_animatingTargetIndex; }
     int GetAnimatingOldIndexOffset() const { return m_animatingOldIndexOffset; }
 
 private:
